+++ src/utils/str/ft_strncmp.c
@@ -1,27 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   str2.c                                             :+:      :+:    :+:   */
+/*   ft_strncmp.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mniemaz <mniemaz@student.42lyon.fr>        +#+  +:+       +#+        */
+/*   By: nle-gued <nle-gued@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/10 10:00:00 by mniemaz           #+#    #+#             */
-/*   Updated: 2025/04/17 14:09:42 by mniemaz          ###   ########.fr       */
+/*   Created: 2025/04/18 16:32:43 by nle-gued          #+#    #+#             */
+/*   Updated: 2025/04/18 16:32:52 by nle-gued         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-int	is_char_in_str(char c, char *str)
-{
-	int	i;
-
-	i = -1;
-	while (str[++i])
-		if (c == str[i])
-			return (1);
-	return (0);
-}
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
@@ -33,20 +22,4 @@ int	ft_strncmp(const char *s1, const char *s2, size_t n)
 	while (i + 1 < n && s1[i] && s2[i] && s1[i] == s2[i])
 		i++;
 	return ((unsigned char)s1[i] - (unsigned char)s2[i]);
-}
-
-int	ft_isalpha(int c)
-{
-	return ((c >= 'a' && c <= 'z') || (c >= 'A' && c <= 'Z'));
-}
-
-int	ft_isalnum(int c)
-{
-	return ((c >= 'a' && c <= 'z') || (c >= 'A' && c <= 'Z') || (c >= '0'
-			&& c <= '9'));
-}
-
-int	ft_isdigit(int c)
-{
-	return (c >= '0' && c <= '9');
 }
